# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/visitRequests": {
            "get": {
                "tags": [
                    "Visit Requests"
                ],
                "description": "All visit requests",
                "operationId": "c4dcea50b63ab8c52fe17b20a026a4ba",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All visit requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "property_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 2
                                                },
                                                "requested_at": {
                                                    "type": "dateTime",
                                                    "example": -1991
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Create a visist request"
                ],
                "description": "create a visit request",
                "operationId": "541f009c7808f1e3b98573eb28c11744",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "property_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "requested_at": {
                                        "type": "dateTime",
                                        "example": "2024-10-23 17:47:12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a visist request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Visit request created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "property_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 2
                                                },
                                                "requested_at": {
                                                    "type": "dateTime",
                                                    "example": -1991
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/visitRequests/{visitRequest}": {
            "patch": {
                "tags": [
                    "Update Visit Request"
                ],
                "description": "Update visit request",
                "operationId": "fd3e64e5445a08dd0c8071f07f09a88f",
                "parameters": [
                    {
                        "name": "visitRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ],
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update visit request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Visit request status updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "property_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "unsignedBigInteger",
                                                    "example": 2
                                                },
                                                "requested_at": {
                                                    "type": "dateTime",
                                                    "example": -1991
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}